pyproject.toml
[project]
name = "ai-agents-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "catboost>=1.2.10",
    "category-encoders>=2.9.0",
    "joblib>=1.5.3",
    "kaggle>=2.0.0",
    "langchain>=1.2.13",
    "langchain-core>=1.2.22",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "optuna>=4.8.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "xgboost>=3.2.0",
    "beautifulsoup4>=4.13.0",
    "faiss-cpu>=1.12.0",
    "httpx>=0.28.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.1.0",
]
